Added Specified Input and Output Relations
The original D-Acts depended on a variety of relations that have to do with ways different documents and roles are specified inputs or outputs in processes. E.g., a deontic act document may be a specified input of some process, and a specified output of that same process is a new deontic role. All relations asserted here are ways to participate.

Important changes were needed for this, however. I needed to convert the respective ranges and domains from "OBI:PlannedProcess" to CCO:"IntentionalAct" (which has new label CCO:'PlannedAct')